Saturday 3:40pm, added user flow
