+++ client/components/JoinParty.js
@@ -1,66 +1,44 @@
-import React from 'react'
-import {connect} from 'react-redux'
-import {joinParty} from '../store/playlistStore'
+import React, {Component} from 'react'
+import {Link} from 'react-router-dom'
+import {Button} from 'semantic-ui-react'
 
-export class JoinParty extends React.Component {
+export default class JoinParty extends Component {
   constructor(props) {
     super(props)
     this.state = {
-      partyCode: '',
-      partyExists: false
+      playlistId: ''
     }
-    this.changeHandler = this.changeHandler.bind(this)
-    this.submitHandler = this.submitHandler.bind(this)
+    this.handleChange = this.handleChange.bind(this)
   }
 
-  async changeHandler(event) {
+  handleChange(event) {
     this.setState({
       [event.target.name]: event.target.value
     })
-    if (this.state.partyCode.length === 22) {
-      const backFromDB = await this.props.findParty(this.state.partyCode)
-      if (backFromDB) {
-        this.setState({partyExists: true})
-      }
-    }
-  }
-
-  submitHandler(evt) {
-    evt.preventdefault()
-    // redirect to the socket room
   }
 
   render() {
     return (
-      <div>
-        <form onSubmit={this.submitHandler}>
-          <label>Party Code</label>
+      <div id="create-playlist">
+        <div />
+        <form onSubmit={this.handleSubmit}>
           <input
-            name="partyCode"
             type="text"
-            value={this.state.partyCode}
-            onChange={this.changeHandler}
-            placeholder="Enter Party Code"
+            className="input"
+            name="playlistId"
+            placeholder="Playlist Code..."
+            onChange={this.handleChange}
           />
-          {this.state.partyExists ? (
-            <button type="submit">Join Party</button>
+          {this.state.playlistId.length === 22 ? (
+            <Link to={`/playlist/${this.state.playlistId}`}>Join Party</Link>
           ) : (
-            <button disabled type="submit">
-              Join Party
-            </button>
+            <p>
+              Please Copy and paste the 22 character Spotify Playlist Id in the
+              search box. we will use this to create your musicVote playlist
+            </p>
           )}
         </form>
       </div>
     )
   }
 }
-
-const mapStateToProps = state => ({
-  partyJoined: state.user.partyIn
-})
-
-const mapDispatchToProps = dispatch => ({
-  findParty: code => dispatch(findThatParty(code))
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(JoinParty)
